Перенесла методы статистики пользователя в модель пользователя
