+++ controllers/ProfileController.php
@@ -3,7 +3,6 @@ namespace app\controllers;
 
 use yii\web\Controller;
 use omarinina\domain\models\user\Users;
-use omarinina\infrastructure\statistic\ExecutorStatistic;
 use yii\web\NotFoundHttpException;
 
 class ProfileController extends Controller
@@ -11,13 +10,12 @@ class ProfileController extends Controller
     /**
      * @param int $id
      * @return string
-     * @throws NotFoundHttpException
+     * @throws NotFoundHttpException|\yii\base\InvalidConfigException
      */
     public function actionView(int $id): string
     {
         if ($id) {
             $currentUser = Users::findOne($id);
-            $executorStatistic = new ExecutorStatistic($currentUser);
             if (!$currentUser || $currentUser->userRole->role !== 'executor') {
                 throw new NotFoundHttpException('Task is not found', 404);
             }
@@ -26,8 +24,7 @@ class ProfileController extends Controller
         }
 
         return $this->render('view', [
-            'currentUser' => $currentUser,
-            'executorStatistic' => $executorStatistic
+            'currentUser' => $currentUser
         ]);
     }
 }
